Extract tests in a dedicated directory
I know this is not the rust way to do things, but I'm just trying to
get comfy with rust for now.
