+++ src/ui/new_renderer.rs
@@ -139,28 +139,3 @@ impl AppComponent {
     }
 }
 
-#[cfg(test)]
-mod tests {
-    use super::*;
-
-    #[tokio::test]
-    async fn test_app_creation() {
-        // This would require a mock sync service
-        // let sync_service = SyncService::new("dummy_token".to_string()).await.unwrap();
-        // let app = AppComponent::new(sync_service);
-        // assert!(!app.should_quit());
-    }
-
-    #[tokio::test]
-    async fn test_event_handling() {
-        // Test that the event handler can be created
-        let event_handler = EventHandler::new();
-
-        // Initially should not need to render (just created)
-        assert!(!event_handler.should_render());
-
-        // After waiting, should be ready to render
-        tokio::time::sleep(tokio::time::Duration::from_millis(17)).await;
-        assert!(event_handler.should_render());
-    }
-}
