+++ tests/date_utils_test.rs
@@ -1,0 +1,24 @@
+use terminalist::utils::date::*;
+use chrono::{NaiveDate, Weekday};
+
+#[test]
+fn test_format_ymd() {
+    let date = NaiveDate::from_ymd_opt(2023, 12, 25).unwrap();
+    assert_eq!(format_ymd(date), "2023-12-25");
+}
+
+#[test]
+fn test_next_weekday_monday() {
+    let friday = NaiveDate::from_ymd_opt(2023, 12, 22).unwrap(); // Friday
+    let next_monday = next_weekday(friday, Weekday::Mon);
+    let expected = NaiveDate::from_ymd_opt(2023, 12, 25).unwrap(); // Next Monday
+    assert_eq!(next_monday, expected);
+}
+
+#[test]
+fn test_next_weekday_same_day() {
+    let monday = NaiveDate::from_ymd_opt(2023, 12, 25).unwrap(); // Monday
+    let next_monday = next_weekday(monday, Weekday::Mon);
+    let expected = NaiveDate::from_ymd_opt(2024, 1, 1).unwrap(); // Next Monday (7 days later)
+    assert_eq!(next_monday, expected);
+}
